Переименование custom_buffer -> ring_buffer (#27)
* refactor: layer_container_t

* refactor: rename custom_buffer.h

* refactor: layer_container -> ring_buffer

* refactor: CMakeLists

---------
